Backend/requirements.txt
supabase
fastapi
fastapi[standard]
python-dotenv
pyjwt
passlib[bcrypt]
pandas
scikit-learn